fix and optimization of code
navigation and search is working
