+++ src/JS/commonFunction.js
@@ -1,5 +1,9 @@
 // масиив жанров
 import genreID from '../JS/genreID';
+import refs from './refs';
+import modalMovieTemplate from '../templates/fullDescriptionMovie.hbs';
+import getMovieFromSaved from './getMovieFromSaved';
+import addMovieToSaved from './addMovieToSaved';
 
 // изменяет  жанр  и дату
 function changeGenreData(filmsData) {
@@ -40,4 +44,72 @@ function changeNumberOfItem(filmsData) {
   return newFilmsData;
 }
 
-export { changeGenreData, changeNumberOfItem };
+function storageModal() {
+  // берем элементы из DOM (доступны после вставки разметки )
+  let filmItemRef = document.querySelectorAll('.trend-film-item');
+  const filmCardRef = document.querySelector('.film-card');
+
+  filmItemRef.forEach(data => {
+    // берем id фильма
+    let filmID = data.dataset.film;
+
+    // Вешаем на картинку слушателя, чтобы открывалась модалка
+    data.addEventListener('click', modalHandler);
+
+    function modalHandler(event) {
+      event.preventDefault();
+
+      fetch(
+        `https://api.themoviedb.org/3/movie/${filmID}?api_key=bf08c0c07642287cbabe383c02818eb3`,
+      )
+        .then(response => response.json())
+        .then(movie => {
+          const markup2 = modalMovieTemplate(movie);
+
+          filmCardRef.insertAdjacentHTML('beforeend', markup2);
+
+          refs.movieModal.classList.remove('is-hidden');
+
+          // додаємо роботу із localStorage
+          // беремо посилання на кнопки
+          const addToWatchedBtnRef = document.querySelector(
+            '.js-add-to-watched',
+          );
+          const addToQueueBtnRef = document.querySelector('.js-add-to-queue');
+
+          // отримуємо масиви фільмів "Watched" та "Queue" із localStorage
+          let watchedMoviesList = getMovieFromSaved('watched');
+          let queueMoviesList = getMovieFromSaved('queue');
+
+          // додаємо фільм у localStorage
+          addMovieToSaved(
+            movie,
+            watchedMoviesList,
+            'watched',
+            addToWatchedBtnRef,
+          );
+          addMovieToSaved(movie, queueMoviesList, 'queue', addToQueueBtnRef);
+
+          function closeModal() {
+            refs.movieModal.classList.add('is-hidden');
+            window.removeEventListener('keydown', pressEscape);
+            filmCardRef.innerHTML = '';
+          }
+
+          const closeBtnRef = document.querySelector('.close-button');
+          closeBtnRef.addEventListener('click', closeModal);
+          const backdropRef = document.querySelector('.backdrop');
+          backdropRef.addEventListener('click', closeModal);
+
+          const pressEscape = event => {
+            if (event.code === 'Escape') {
+              closeModal();
+            }
+          };
+          window.addEventListener('keydown', pressEscape);
+        });
+    }
+  });
+}
+
+export { changeGenreData, changeNumberOfItem, storageModal };
