+++ src/JS/searchResultsMarkup.js
@@ -1,46 +1,24 @@
 import film from '../templates/trendMovieTemplate.hbs';
 import refs from './refs';
-import genreID from '../JS/genreID';
+
+import {
+  changeGenreData,
+  changeNumberOfItem,
+  storageModal,
+} from './commonFunction';
 
 // --------выводим результат поиска  по ключевым словам------
 const searchResultsMarkup = results => {
   // изменяет  жанр  и дату
-  results.map(item => {
-    let newGenre = [];
-    item.genre_ids.map(id => {
-      const found = genreID.find(item => item.id === id);
-      newGenre.push(found.name);
-    });
-    item.release_date = item.release_date.slice(0, 4);
-    if (newGenre.length >= 4) {
-      const manyGenres = newGenre.slice(0, 3);
-      item.genre_ids = manyGenres.join(', ');
-    } else {
-      item.genre_ids = newGenre.join(', ');
-    }
-    return item;
-  });
+  changeGenreData(results);
 
-  let newMovieList = [];
+  // в зависимости от viewport оствляет необходимое количество элементов
+  let newFilmCount = changeNumberOfItem(results);
 
-  if (document.documentElement.clientWidth < 768) {
-    newMovieList = results.slice(0, 4);
-    // return  newMovieList
-  }
-  if (
-    document.documentElement.clientWidth >= 768 &&
-    document.documentElement.clientWidth < 1024
-  ) {
-    newMovieList = results.slice(0, 8);
-  }
-  if (document.documentElement.clientWidth >= 1024) {
-    newMovieList = results.slice(0, 9);
-  }
+  const markup = film(newFilmCount);
+  refs.movies.insertAdjacentHTML('beforeend', markup);
 
-  console.log(newMovieList);
-
-  const markup = film(newMovieList);
-  refs.movieRef.insertAdjacentHTML('beforeend', markup);
+  storageModal();
 
   // ------------условие для вывода предупреждения---------------
   if (results === []) {
