+++ src/scss/layout/_footer.scss
@@ -61,7 +61,7 @@ footer > div {
   left: 0;
   width: 100%;
   height: 100%;
-  // background-color: rgba(0, 0, 0, 0.5);
+  background-color: rgba(0, 0, 0, 0.5);
   z-index: 20;
   transition: 0.3s all;
 }
