+++ src/scss/layout/_trendMovee.scss
@@ -19,7 +19,7 @@
     width: calc((100% - 4 * 15px) / 2);
   }
   @media screen and (min-width: 1024px) {
-    width: calc((100% - 6 * 15px) / 3);
+    width: calc((100% - 12 * 15px) / 3);
   }
 }
 
@@ -29,8 +29,6 @@
 }
 .trend-film-name {
   margin: 0px;
-  //     font-family: Roboto;
-  // font-weight: 500;
   font-size: 12px;
   line-height: 1.14;
   text-transform: uppercase;
